renamed files and added a empty main with require for different files
